use appropriate VectorStats/optimize_interval and fix change detector test
